move curve/curves_to_focus to coil/coils_to_focus
