+++ src/simsopt/geo/curve.py
@@ -846,93 +846,3 @@ def curves_to_vtk(curves, filename, close=False):
     data = np.concatenate([i*np.ones((ppl[i], )) for i in range(len(curves))])
     polyLinesToVTK(filename, x, y, z, pointsPerLine=ppl, pointData={'idx': data})
 
-
-def curves_to_focus(filename, curves, currents, nfp=1, stellsym=False, Ifree=False, Lfree=False):
-    """
-    Export a list of Curve objects together with currents in FOCUS format, so they can 
-    be used by FOCUS. The format is introduced at
-    https://princetonuniversity.github.io/FOCUS/rdcoils.pdf
-
-    Args:
-        filename: Name of the file to write.
-        curves: A python list of Curve objects.
-        currents: Coil current of each curve.
-        nfp: The number of field periodicity. Defaults to 1.      
-        stellsym: Whether or not following stellarator symmetry. Defaults to False.
-        Ifree: Flag specifying whether the coil current is free. Defaults to False.
-        Lfree: Flag specifying whether the coil geometry is free. Defaults to False.
-    """
-    from simsopt.geo import CurveLength
-
-    assert len(curves) == len(currents)
-    ncoils = len(curves)
-    if stellsym:
-        symm = 2  # both periodic and symmetric
-    elif nfp > 1 and not stellsym:
-        symm = 1  # only periodicity
-    else:
-        symm = 0  # no periodicity or symmetry
-    if nfp > 1:
-        print('Please note: FOCUS sets Nfp in the plasma file.')
-    with open(filename, 'w') as f:
-        f.write('# Total number of coils \n')
-        f.write('  {:d} \n'.format(ncoils))
-        for i in range(ncoils):
-            nf = curves[i].order
-            xyz = curves[i].full_x.reshape((3, -1))
-            xc = xyz[0, ::2]
-            xs = np.concatenate(([0.], xyz[0, 1::2]))
-            yc = xyz[1, ::2]
-            ys = np.concatenate(([0.], xyz[1, 1::2]))
-            zc = xyz[2, ::2]
-            zs = np.concatenate(([0.], xyz[2, 1::2]))
-            length = CurveLength(curves[i]).J()
-            nseg = len(curves[i].quadpoints)
-            f.write('#------------{:d}----------- \n'.format(i+1))
-            f.write('# coil_type  symm  coil_name \n')
-            f.write('  {:d}   {:d}  {:} \n'.format(1, symm, curves[i].name))
-            f.write('# Nseg current Ifree Length Lfree target_length \n')
-            f.write('  {:d} {:23.15E} {:d} {:23.15E} {:d} {:23.15E} \n'.format(nseg, currents[i].get_value(), Ifree, length, Lfree, length))
-            f.write('# NFcoil \n')
-            f.write('  {:d} \n'.format(nf))
-            f.write('# Fourier harmonics for coils ( xc; xs; yc; ys; zc; zs) \n')
-            for r in [xc, xs, yc, ys, zc, zs]:  # 6 lines
-                for k in range(nf+1):
-                    f.write('{:23.15E} '.format(r[k]))
-                f.write('\n')
-        f.write('\n')
-    return
-
-
-def create_equally_spaced_curves(ncurves, nfp, stellsym, R0=1.0, R1=0.5, order=6, numquadpoints=None):
-    """
-    Create ``ncurves`` curves of type
-    :obj:`~simsopt.geo.curvexyzfourier.CurveXYZFourier` of order
-    ``order`` that will result in circular equally spaced coils (major
-    radius ``R0`` and minor radius ``R1``) after applying
-    :obj:`~simsopt.field.coil.coils_via_symmetries`.
-
-    Usage example: create 4 base curves, which are then rotated 3 times and
-    flipped for stellarator symmetry:
-
-    .. code-block::
-
-        base_curves = create_equally_spaced_curves(4, 3, stellsym=True)
-        base_currents = [Current(1e5) for c in base_curves]
-        coils = coils_via_symmetries(base_curves, base_currents, 3, stellsym=True)
-    """
-    if numquadpoints is None:
-        numquadpoints = 15 * order
-    curves = []
-    from simsopt.geo.curvexyzfourier import CurveXYZFourier
-    for i in range(ncurves):
-        curve = CurveXYZFourier(numquadpoints, order)
-        angle = (i+0.5)*(2*np.pi)/((1+int(stellsym))*nfp*ncurves)
-        curve.set("xc(0)", cos(angle)*R0)
-        curve.set("xc(1)", cos(angle)*R1)
-        curve.set("yc(0)", sin(angle)*R0)
-        curve.set("yc(1)", sin(angle)*R1)
-        curve.set("zs(1)", R1)
-        curve.x = curve.x  # need to do this to transfer data to C++
-        curves.append(curve)
-    return curves
